+++ app/models/manageiq/providers/redhat/infra_manager/event_parser.rb
@@ -34,9 +34,7 @@ module ManageIQ::Providers::Redhat::InfraManager::EventParser
     user_href = ems_ref_from_object_in_event(event[:user])
     username  = nil
     if ems && user_href
-      ems.with_provider_connection do |rhevm|
-        username = Ovirt::User.find_by_href(rhevm, user_href).try(:[], :user_name)
-      end
+      username = ems.inventory.collect_username_by_href(ems, user_href)
     end
 
     # Build the event hash
@@ -63,9 +61,7 @@ module ManageIQ::Providers::Redhat::InfraManager::EventParser
     cluster = full_data[:cluster]
     cluster_ref = ManageIQ::Providers::Redhat::InfraManager.make_ems_ref(cluster[:href])
 
-    cluster_name = ems.with_provider_connection do |rhevm|
-      Ovirt::Cluster.find_by_href(rhevm, cluster_ref).try(:[], :name)
-    end
+    cluster_name = ems.inventory.collect_cluster_name_href(ems, cluster_ref)
 
     {
       :ems_id         => ems.id,
