+++ spec/models/manageiq/providers/redhat/infra_manager/event_parser/response_yamls/user.yml
@@ -1,0 +1,41 @@
+--- !ruby/object:OvirtSDK4::User
+href: "/ovirt-engine/api/users/58ad9d2d-013a-00aa-018f-00000000022e"
+comment: 
+description: 
+id: 58ad9d2d-013a-00aa-018f-00000000022e
+name: admin
+department: 
+domain: !ruby/object:OvirtSDK4::Domain
+  href: "/ovirt-engine/api/domains/696E7465726E616C2D617574687A"
+  comment: 
+  description: 
+  id: 696E7465726E616C2D617574687A
+  name: internal-authz
+  groups: 
+  user: 
+  users: 
+domain_entry_id: 30663066383031392D353864322D343937302D626464302D666561323334613732663263
+email: 
+groups: 
+last_name: 
+logged_in: 
+namespace: "*"
+password: 
+permissions: !ruby/array:OvirtSDK4::List
+  internal: []
+  ivars:
+    :@href: "/ovirt-engine/api/users/58ad9d2d-013a-00aa-018f-00000000022e/permissions"
+principal: admin
+roles: !ruby/array:OvirtSDK4::List
+  internal: []
+  ivars:
+    :@href: "/ovirt-engine/api/users/58ad9d2d-013a-00aa-018f-00000000022e/roles"
+ssh_public_keys: !ruby/array:OvirtSDK4::List
+  internal: []
+  ivars:
+    :@href: "/ovirt-engine/api/users/58ad9d2d-013a-00aa-018f-00000000022e/sshpublickeys"
+tags: !ruby/array:OvirtSDK4::List
+  internal: []
+  ivars:
+    :@href: "/ovirt-engine/api/users/58ad9d2d-013a-00aa-018f-00000000022e/tags"
+user_name: admin@internal-authz
