+++ spec/models/manageiq/providers/redhat/infra_manager/refresh/refresher/response_yamls/datacenters.yml
@@ -1,5 +1,5 @@
 ---
-- !ruby/marshalable:ManageIQ::Providers::Redhat::InfraManager::Refresh::Strategies::DatacenterPreloadedAttributesDecorator
+- !ruby/marshalable:ManageIQ::Providers::Redhat::InfraManager::Inventory::DatacenterPreloadedAttributesDecorator
   :__v2__:
   - :@obj
   - :@storage_domains
