+++ spec/models/manageiq/providers/redhat/infra_manager/refresh/refresher/response_yamls/hosts.yml
@@ -1,5 +1,5 @@
 ---
-- !ruby/marshalable:ManageIQ::Providers::Redhat::InfraManager::Refresh::Strategies::HostPreloadedAttributesDecorator
+- !ruby/marshalable:ManageIQ::Providers::Redhat::InfraManager::Inventory::HostPreloadedAttributesDecorator
   :__v2__:
   - :@obj
   - :@nics
