+++ spec/models/manageiq/providers/redhat/infra_manager/refresh/refresher_4_1_spec.rb
@@ -18,7 +18,7 @@ describe ManageIQ::Providers::Redhat::InfraManager::Refresh::Refresher do
   end
 
   before(:each) do
-    inventory_wrapper_class = ManageIQ::Providers::Redhat::InfraManager::Refresh::Strategies::Api4::InventoryWrapper
+    inventory_wrapper_class = ManageIQ::Providers::Redhat::InfraManager::Inventory::InventoryV4
     allow_any_instance_of(inventory_wrapper_class)
       .to receive(:collect_clusters).and_return(load_response_mock_for('clusters'))
     allow_any_instance_of(inventory_wrapper_class)
